refactor(three-tier): retire the untiered alttprbot/tournament/ package
Delete the last untiered god-object package. Rather than rewrite the live
RaceTime/cog dispatch, relocate the transitional bridge (the TournamentRace-shaped
OrchestratorAdapter that every active event slug already resolves to) into the tier
where it belongs, and drop the dead remnants.

Relocations (behavior-preserving):
- orchestrator_adapter.py -> presentation/discord/tournament/ (it touches discordbot +
  the presenter + holds the live RaceTime handler). Player lookup now goes through
  UserService instead of UserRepository (identical delegation; presentation->service).
- registry_loader.py -> services/tournament/ (byte-identical; import-pure config logic).
- TournamentConfig dataclass -> presentation/discord/tournament/config.py (the resolved,
  discord-typed counterpart of the service-tier TournamentDefinition).
- UnableToLookupUserException -> alttprbot/exceptions.py (stays a SahasrahBotException).

Dead code deleted:
- core.py's TournamentRace base + TournamentPlayer class (never instantiated; the adapter
  uses services/tournament/types.TournamentPlayer) and the entire alttpr.py
  (ALTTPRTournamentRace / ALTTPR2024Race, never instantiated or subclassed by live code).

Cog cleanup (presentation/discord/cogs/tournament.py):
- Remove the cc2023 and tournament_deck slash commands. They called
  alttpr.roll_seed / alttpr.generate_deck, which have never existed in any branch
  (including master) — i.e. they have been broken (AttributeError on invocation) since
  before the migration. Removed per owner decision, with their stale commented-out
  experimental blocks and the now-unused random/seed_embed/core/alttpr imports and the
  core.* type annotations. The ChallengeCup persistent views + their on_ready
  registration are retained (existing Discord messages depend on them).

Importers updated: tournaments.py (untiered dispatch glue, retained), helpers/, and the
adapter test. Behavior-preserving: 462 tests green, import-linter 3 kept / 0 broken, all
surfaces import cleanly, and the registry builds the expected production slugs. Reviewed
adversarially (dead-code-deletion safety, cog removal, relocation parity).
