+++ alttprbot/presentation/discord/cogs/tournament.py
@@ -1,6 +1,5 @@
 import datetime
 import logging
-import random
 import asyncio
 
 import discord
@@ -8,11 +7,8 @@ from discord import app_commands
 from discord.ext import commands, tasks
 
 import config
-# from alttprbot.presentation.discord.util import alttpr_discord
 from alttprbot.services import TournamentSchedulingService
 from alttprbot import tournaments
-from alttprbot.presentation.discord.util.seed_embeds import seed_embed
-from alttprbot.tournament import core, alttpr
 from alttprbot.util import speedgaming
 
 # TODO: use asyncio.semaphore() to limit the number of concurrent tasks
@@ -83,7 +79,7 @@ class Tournament(commands.Cog):
         try:
             logging.info("scanning SG schedule for tournament races to create")
             for event_slug, tournament_class in tournaments.TOURNAMENT_DATA.items():
-                event_data: core.TournamentConfig = await tournament_class.get_config()
+                event_data = await tournament_class.get_config()
                 try:
                     episodes = await speedgaming.get_upcoming_episodes_by_event(event_slug, hours_past=0.5,
                                                                                 hours_future=event_data.hours_before_room_open)
@@ -102,7 +98,7 @@ class Tournament(commands.Cog):
         try:
             logging.info('scanning for unsubmitted races')
             for event_slug, tournament_class in tournaments.TOURNAMENT_DATA.items():
-                event_data: core.TournamentRace = await tournament_class.get_config()
+                event_data = await tournament_class.get_config()
 
                 try:
                     episodes = await speedgaming.get_upcoming_episodes_by_event(event_slug, hours_past=0,
@@ -130,7 +126,7 @@ class Tournament(commands.Cog):
             for event_slug, tournament_class in tournaments.TOURNAMENT_DATA.items():
                 try:
                     messages = await self.report_bad_player_discord(event_slug=event_slug)
-                    event_data: core.TournamentConfig = await tournament_class.get_config()
+                    event_data = await tournament_class.get_config()
                 except Exception:
                     logging.exception("Error while checking bad discord data for event %s", event_slug)
                     continue
@@ -186,7 +182,7 @@ class Tournament(commands.Cog):
         else:
             return False
 
-    async def update_scheduled_event(self, event_data: core.TournamentRace, event_slug: str, episodes: dict):
+    async def update_scheduled_event(self, event_data, event_slug: str, episodes: dict):
 
         # remove dead events
         scheduling_service = TournamentSchedulingService()
@@ -277,7 +273,7 @@ class Tournament(commands.Cog):
             except Exception:
                 logging.exception("Unable to create guild event.")
 
-    async def update_scheduling_needs(self, event_data: core.TournamentRace, episodes):
+    async def update_scheduling_needs(self, event_data, episodes):
         comms_needed = []
         trackers_needed = []
         broadcasters_needed = []
@@ -379,7 +375,7 @@ class Tournament(commands.Cog):
     async def report_bad_player_discord(self, event_slug):
         tournament_class = tournaments.TOURNAMENT_DATA[event_slug]
 
-        event_data: core.TournamentConfig = await tournament_class.get_config()
+        event_data = await tournament_class.get_config()
         episodes = await speedgaming.get_upcoming_episodes_by_event(event_slug, hours_past=0, hours_future=48)
 
         messages = []
@@ -416,140 +412,6 @@ class Tournament(commands.Cog):
                                 f"Episode {episode['id']} - {event_slug} - {player['displayName']} could not be found")
 
         return messages
-
-    # @app_commands.command(description="Generate an ALTTPR practice seed from an SG Episode that's already been submitted.")
-    # async def practice(self, interaction: discord.Interaction, episode_id: int):
-    #     await interaction.response.defer()
-    #     tournament_game = await models.TournamentGames.get_or_none(episode_id=episode_id)
-    #     if tournament_game is None:
-    #         await interaction.response.send_message("That episode has not been submitted yet.", ephemeral=True)
-    #         return
-
-    #     settings = tournament_game.settings
-
-    #     seed = await alttpr_discord.ALTTPRDiscord.generate(settings=settings, endpoint='/api/customizer')  # TODO: don't hardcode endpoint
-    #     embed = await seed.embed(emojis=self.bot.emojis)
-
-    #     await interaction.followup.send(embed=embed)
-
-    # @app_commands.command(description="Generate a randomizer seed for the ALTTPR Main Tournament 2024.")
-    # async def 4(self, interaction: discord.Interaction, player1: discord.Member, player2: discord.Member):
-    #     await interaction.response.defer()
-    #     seed, preset, deck = await alttpr.roll_seed([player1, player2])
-
-    #     embed = await seed.embed(emojis=self.bot.emojis)
-    #     embed.insert_field_at(0, name="Preset", value=preset, inline=False)
-    #     if deck:
-    #         embed.insert_field_at(1, name="Deck", value="\n".join([f"**{p}**: {c}" for p, c in deck.items()]), inline=False)
-
-    #     await interaction.response.send_message(embed=embed)
-
-    @app_commands.command(description="Generate a randomizer seed for the 2023 challenge cup.")
-    @app_commands.guilds(*CC_TOURNAMENT_SERVERS)
-    async def cc2023(self, interaction: discord.Interaction, opponent: discord.Member,
-                     on_behalf_of: discord.Member = None, player3: discord.Member = None,
-                     player4: discord.Member = None):
-        if interaction.guild.chunked is False:
-            await interaction.guild.chunk(cache=True)
-
-        if on_behalf_of and interaction.guild.get_role(CC_TOURNAMENT_ADMIN_ROLE_ID) not in interaction.user.roles:
-            await interaction.response.send_message(
-                "You must be a member of the Challenge Cup admin team to roll a seed on someone else's behalf..")
-            return
-
-        if (player3 or player4) and interaction.guild.get_role(
-                CC_TOURNAMENT_ADMIN_ROLE_ID) not in interaction.user.roles:
-            await interaction.response.send_message(
-                "You must be a member of the Challenge Cup admin team to roll a seed with more than 2 players.")
-            return
-
-        if interaction.user == opponent:
-            await interaction.response.send_message("You can't race yourself.", ephemeral=True)
-            return
-
-        if on_behalf_of is None:
-            on_behalf_of = interaction.user
-
-        if opponent.bot or on_behalf_of.bot:
-            await interaction.response.send_message("You can't race a bot.", ephemeral=True)
-            return
-
-        players = [opponent, on_behalf_of]
-        if player3:
-            players.append(player3)
-        if player4:
-            players.append(player4)
-
-        await interaction.response.defer()
-        # embed = await self.generate_deck_seed(players, "cc2023")
-
-        msg_id = None
-
-        if CC_TOURNAMENT_AUDIT_CHANNELS:
-            channel = self.bot.get_channel(CC_TOURNAMENT_AUDIT_CHANNELS)
-            msg = await channel.send("Generating...")
-            msg_id = msg.id
-
-        seed, preset, deck = await alttpr.roll_seed(players, event_slug="cc2023", episode_id=msg_id)
-
-        embed = await seed_embed(seed, emojis=self.bot.emojis, include_settings=False)
-        embed.insert_field_at(0, name="Preset", value=preset, inline=False)
-        if deck:
-            embed.insert_field_at(1, name="Deck", value="\n".join([f"**{p}**: {c}" for p, c in deck.items()]),
-                                  inline=False)
-
-        embed.title = " vs. ".join([p.display_name for p in players])
-        embed.description = " vs. ".join([p.mention for p in players])
-
-        for player in players:
-            await player.send(embed=embed)
-
-        if CC_TOURNAMENT_AUDIT_CHANNELS:
-            await msg.edit(content="Sent to players.", embed=embed, view=ChallengeCupDeleteHistoryView())
-
-        await interaction.followup.send("Seed successfully sent to DM.")
-
-    @app_commands.command(
-        description="Generate a hypothetical deck for a match.  This does not generate a seed or write to history.")
-    @app_commands.guilds(*CC_TOURNAMENT_SERVERS, *MAIN_TOURNAMENT_SERVERS)
-    @app_commands.choices(
-        event_slug=[
-            app_commands.Choice(name="cc2024", value="cc2024"),
-            app_commands.Choice(name="alttpr2024", value="alttpr2024"),
-        ]
-    )
-    async def tournament_deck(self, interaction: discord.Interaction, opponent: discord.Member,
-                              on_behalf_of: discord.Member = None, event_slug: str = None):
-        if on_behalf_of is None:
-            on_behalf_of = interaction.user
-
-        if interaction.user == opponent:
-            await interaction.response.send_message("You must specify two different players.", ephemeral=True)
-            return
-
-        if opponent.bot or on_behalf_of.bot:
-            await interaction.response.send_message("You cannot specify a bot as a player.", ephemeral=True)
-            return
-
-        if event_slug is None:
-            if interaction.guild.id in CC_TOURNAMENT_SERVERS:
-                event_slug = "cc2023"
-            elif interaction.guild.id in MAIN_TOURNAMENT_SERVERS:
-                event_slug = "alttpr2024"
-            else:
-                await interaction.response.send_message("You must specify an event slug.", ephemeral=True)
-                return
-
-        deck = await alttpr.generate_deck([opponent, on_behalf_of], event_slug=event_slug)
-        preset = random.choices(list(deck.keys()), weights=list(deck.values()))[0]
-        embed = discord.Embed(
-            title=f"{opponent.display_name} vs. {on_behalf_of.display_name}",
-            description=f"{opponent.mention} vs. {on_behalf_of.mention}",
-            color=discord.Color.blue()
-        )
-        embed.add_field(name="Deck", value="\n".join([f"**{p}**: {c}" for p, c in deck.items()]), inline=False)
-        embed.add_field(name="In this hypothetical matchup, this mode was drawn:", value=preset, inline=False)
-        await interaction.response.send_message(embed=embed, ephemeral=True)
 
 
 async def setup(bot: commands.Bot):
