+++ alttprbot/services/tournament/registry_loader.py
